Move tap key for right alt from backslash to right bracket.
Since backslash is used as leader key in Vim, and I use this key a lot
for a variety of key mappings, sometimes it would be accidentally
treated as right alt if I type key mapping beginning with backslash
too fast.

To avoid such problem, I move tap key to right bracket, which is used
less less frequently.